backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
langchain==0.3.19
langchain-groq==0.3.1
sqlalchemy==2.0.35
alembic==1.15.1
python-dotenv==1.0.1
pydantic-settings==2.5.2
python-multipart==0.0.9
httpx==0.27.2